[add] further protection / short-circuits
-calling 'll' which sets the selected location list item (/line) is
problematic when a window displaying a buffer with an associated
location list then has a buffer with no associated location list
entries and no mechanism is used to reset said location list. it
appears 'Syntastic' does not automatically clear existing entries
whereas 'Ale' does. when this happens, the call to 'getloclist'
returns what are effectively invalid items (i.e. not applicable to
the current buffer) which are then used by the 'll' call which is
implemented to not only ensure the location list window is made
visible, but also the appropriate buffer too (thus the previous buffer
in this case) causing our intended buffer to immediately get unloaded