[fix] inconsistent state on reload
-force state synchronisation on reload
-script variables lost their state on reload and functionality
continued unaware that it had reverted to using defaults. as no
explicit resetting of enabled buffers' local states is performed,
even re-reading a buffer was inadequate for triggering the
necessary global -> script synchronisation. although inconvenient,
toggling a buffer off/on would have worked, as would the toggling
of the settings themselves